refactor: Separate header and source file
