+++ Includes/Rosetta/Actions/ActionApplyHelper.hpp
@@ -100,53 +100,17 @@ class ActionApplyHelper
     {
      public:
         ActionParamsChoser(const std::vector<int>& choices,
-                           std::size_t& choicesIdx)
-            : m_choices(choices), m_choicesIdx(choicesIdx), m_shouldSkip(false)
-        {
-            // Do nothing
-        }
+                           std::size_t& choicesIdx);
 
-        bool HasFurtherChoices() const
-        {
-            return m_choicesIdx <= m_choices.size();
-        }
+        bool HasFurtherChoices() const;
 
-        bool ShouldSkip() const
-        {
-            return m_shouldSkip;
-        }
+        bool ShouldSkip() const;
 
         std::size_t GetNumber(ActionType actionType,
-                              ActionChoices& actionChoices) final
-        {
-            m_shouldSkip = false;
-
-            if (actionType != ActionType::MAIN_ACTION)
-            {
-                if (actionChoices.Size() == 1)
-                {
-                    m_shouldSkip = true;
-                    return 0;
-                }
-            }
-
-            return GetNextChoice(0, actionChoices.Size());
-        }
+                              ActionChoices& actionChoices) final;
 
      private:
-        int GetNextChoice(int min, [[maybe_unused]] int max)
-        {
-            if (m_choicesIdx >= m_choices.size())
-            {
-                ++m_choicesIdx;
-                return min;
-            }
-
-            const int choice = m_choices[m_choicesIdx];
-            ++m_choicesIdx;
-
-            return choice;
-        }
+        int GetNextChoice(int min, [[maybe_unused]] int max) const;
 
         const std::vector<int>& m_choices;
         std::size_t& m_choicesIdx;
@@ -157,130 +121,37 @@ class ActionApplyHelper
     {
      public:
         ActionParamsCallback(const std::vector<int>& choices,
-                             std::size_t& choicesIdx, CallbackInfo& result)
-            : m_result(result), m_choser(choices, choicesIdx)
-        {
-        }
+                             std::size_t& choicesIdx, CallbackInfo& result);
 
-        void Initialize(const Game& game) override
-        {
-            m_result = NullInfo();
-            m_choser.Initialize(game);
-        }
+        void Initialize(const Game& game) override;
 
-        MainOpType ChooseMainOp() override
-        {
-            auto ret = m_choser.ChooseMainOp();
+        MainOpType ChooseMainOp() override;
 
-            if (ShouldRecord())
-            {
-                m_result = MainOpInfo();
-            }
+        Entity* ChooseHandCard() override;
 
-            return ret;
-        }
+        Character* GetAttacker() override;
 
-        Entity* ChooseHandCard() override
-        {
-            auto ret = m_choser.ChooseHandCard();
-
-            if (ShouldRecord())
-            {
-                m_result = ChooseHandCardInfo();
-            }
-
-            return ret;
-        }
-
-        Character* GetAttacker() override
-        {
-            auto ret = m_choser.GetAttacker();
-
-            if (ShouldRecord())
-            {
-                m_result = ChooseAttackerInfo();
-            }
-
-            return ret;
-        }
-
-        std::size_t GetMinionPutLocation(int minions) override
-        {
-            auto ret = m_choser.GetMinionPutLocation(minions);
-
-            if (ShouldRecord())
-            {
-                m_result = MinionPutLocationInfo(minions);
-            }
-
-            return ret;
-        }
+        std::size_t GetMinionPutLocation(int minions) override;
 
         Character* GetSpecifiedTarget(
-            const std::vector<Character*>& targets) override
-        {
-            auto ret = m_choser.GetSpecifiedTarget(targets);
+            const std::vector<Character*>& targets) override;
 
-            if (ShouldRecord())
-            {
-                m_result = GetSpecifiedTargetInfo(targets);
-            }
-
-            return ret;
-        }
-
-        std::size_t ChooseOne(const std::vector<std::size_t>& cards) override
-        {
-            auto ret = m_choser.ChooseOne(cards);
-
-            if (ShouldRecord())
-            {
-                m_result = ChooseOneInfo(cards);
-            }
-
-            return ret;
-        }
+        std::size_t ChooseOne(const std::vector<std::size_t>& cards) override;
 
         std::size_t GetNumber([[maybe_unused]] ActionType actionType,
-                              [[maybe_unused]] ActionChoices& choices) override
-        {
-            // Do nothing
-            return 0;
-        }
+                              [[maybe_unused]] ActionChoices& choices) override;
 
      private:
-        bool ShouldRecord() const
-        {
-            if (!std::holds_alternative<NullInfo>(m_result))
-            {
-                return false;
-            }
-            if (m_choser.ShouldSkip())
-            {
-                return false;
-            }
-            if (m_choser.HasFurtherChoices())
-            {
-                return false;
-            }
-
-            return true;
-        }
+        bool ShouldRecord() const;
 
         CallbackInfo& m_result;
         ActionParamsChoser m_choser;
     };
 
  public:
-    void AppendChoice(int choice)
-    {
-        m_choices.push_back(choice);
-    }
+    void AppendChoice(int choice);
 
-    void ClearChoices()
-    {
-        m_choices.clear();
-    }
+    void ClearChoices();
 
     template <class StartBoardGetter>
     CallbackInfo ApplyChoices(StartBoardGetter&& startBoardGetter) const
@@ -289,26 +160,9 @@ class ActionApplyHelper
         return ApplyChoices(game);
     }
 
-    CallbackInfo ApplyChoices(Game& game) const
-    {
-        PlayState result;
-        return ApplyChoices(game, result);
-    }
+    CallbackInfo ApplyChoices(Game& game) const;
 
-    CallbackInfo ApplyChoices(Game& game, PlayState& result) const
-    {
-        CallbackInfo info = NullInfo();
-        size_t choicesIdx = 0;
-        ActionParamsCallback actionCallback(m_choices, choicesIdx, info);
-
-        while (m_choices.size() > choicesIdx)
-        {
-            actionCallback.Initialize(game);
-            result = game.PerformAction(actionCallback);
-        }
-
-        return info;
-    }
+    CallbackInfo ApplyChoices(Game& game, PlayState& result) const;
 
  private:
     std::vector<int> m_choices;
