+++ Includes/Rosetta/Views/Types/CardInfo.hpp
@@ -12,8 +12,6 @@
 
 #include <Rosetta/Views/Types/UnknownCards.hpp>
 
-#include <utility>
-
 namespace RosettaStone::Views::Types
 {
 //!
@@ -25,37 +23,18 @@ struct CardInfo
 {
     //! Sets that the card as revealed.
     //! \param _cardID The card ID to set.
-    void SetAsRevealedCard(std::string _cardID)
-    {
-        cardID = std::move(_cardID);
-    }
+    void SetAsRevealedCard(std::string _cardID);
 
     //! Sets that the card as hidden.
     //! \param setID The ID of the card set.
     //! \param cardIdx The card index of the card set.
-    void SetAsHiddenCard(size_t setID, size_t cardIdx)
-    {
-        cardID = INVALID_CARD_ID;
-        unknownCardsSetID = setID;
-        unknownCardsSetCardIdx = cardIdx;
-    }
+    void SetAsHiddenCard(size_t setID, size_t cardIdx);
 
     //! Returns the card ID considering whether it is valid.
     //! \param unknownCardsManager The manager of unknown cards.
     //! \return the card ID considering whether it is valid.
     std::string GetCardID(
-        const UnknownCardsSetsManager& unknownCardsManager) const
-    {
-        if (cardID != INVALID_CARD_ID)
-        {
-            return cardID;
-        }
-        else
-        {
-            return unknownCardsManager.GetCardID(unknownCardsSetID,
-                                                 unknownCardsSetCardIdx);
-        }
-    }
+        const UnknownCardsSetsManager& unknownCardsManager) const;
 
     std::string cardID;
     std::size_t unknownCardsSetID = 0;
